fix streaming
