+++ style/base.css
@@ -226,6 +226,18 @@
   gap: 8px;
 }
 
+/* Streaming response styles */
+.mcp-streaming-response {
+  position: relative;
+  margin-bottom: 16px;
+  padding: 8px 12px;
+  border-radius: 8px;
+  background: var(--jp-layout-color3);
+  margin-right: auto;
+  max-width: 85%;
+  overflow: visible;
+}
+
 .mcp-chat-item {
   padding: 12px;
   border-radius: 8px;
